Replace Pipfile and setuptools with uv
It's been a minute since I've touched this code. I'm going to bring it up to
something like 2025 Python standards, and the first step is to switch to a
pyproject.toml project definition. I'm using uv as project and package manager.

It doesn't look like I was actually using the runtime dependencies in the old
project definition(s). I'm also not going to use any of the same dev
dependencies (like flake8 or nose), so I didn't carry any of the old dependency
information over.

For now, I've added dev dependencies on ruff and ty, which I'll get set up in
subsequent commits. I'll also have to a test runner -- probably pytest, though
mdtest looks interesting -- and convert the existing tests over.

I also set the Python baseline at >=3.13, which is a bit of a jump from the
previous >=3.7. I haven't converted any code to take advantage of the newer
baseline yet. I'm hoping I can get ruff to do some of that for me automatically.